+++ lite/tests/classification/test_accuracy.py
@@ -1,5 +1,5 @@
 import numpy as np
-from valor_lite.cv.classification import (
+from valor_lite.classification import (
     Classification,
     DataLoader,
     MetricType,
@@ -7,13 +7,13 @@ from valor_lite.cv.classification import (
 )
 
 
-def test_precision_computation():
+def test_accuracy_computation():
 
-    # groundtruth, prediction, score, hardmax
+    # groundtruth, prediction, score
     data = np.array(
         [
             # datum 0
-            [0, 0, 0, 1.0, 0],  # tp
+            [0, 0, 0, 1.0, 1],  # tp
             [0, 0, 1, 0.0, 0],  # tn
             [0, 0, 2, 0.0, 0],  # tn
             [0, 0, 3, 0.0, 0],  # tn
@@ -44,7 +44,7 @@ def test_precision_computation():
 
     score_thresholds = np.array([0.25, 0.75], dtype=np.float64)
 
-    (_, precision, _, _, _, _, _) = compute_metrics(
+    (_, _, _, accuracy, _, _, _) = compute_metrics(
         data=data,
         label_metadata=label_metadata,
         score_thresholds=score_thresholds,
@@ -53,21 +53,21 @@ def test_precision_computation():
     )
 
     # score threshold, label, count metric
-    assert precision.shape == (2, 4)
+    assert accuracy.shape == (2, 4)
 
     # score >= 0.25
-    assert precision[0][0] == 1.0
-    assert precision[0][1] == 0.0
-    assert precision[0][2] == 0.0
-    assert precision[0][3] == 1.0
+    assert accuracy[0][0] == 2 / 3
+    assert accuracy[0][1] == 1.0
+    assert accuracy[0][2] == 2 / 3
+    assert accuracy[0][3] == 1.0
     # score >= 0.75
-    assert precision[1][0] == 1.0
-    assert precision[1][1] == 0.0
-    assert precision[1][2] == 0.0
-    assert precision[1][3] == 0.0
+    assert accuracy[1][0] == 2 / 3
+    assert accuracy[1][1] == 1.0
+    assert accuracy[1][2] == 2 / 3
+    assert accuracy[1][3] == 2 / 3
 
 
-def test_precision_basic(basic_classifications: list[Classification]):
+def test_accuracy_basic(basic_classifications: list[Classification]):
     loader = DataLoader()
     loader.add_data(basic_classifications)
     evaluator = loader.finalize()
@@ -81,16 +81,13 @@ def test_precision_basic(basic_classifications: list[Classification]):
         "missing_prediction_labels": [],
     }
 
-    metrics = evaluator.evaluate(
-        score_thresholds=[0.25, 0.75],
-        as_dict=True,
-    )
+    metrics = evaluator.evaluate(score_thresholds=[0.25, 0.75], as_dict=True)
 
-    actual_metrics = [m for m in metrics[MetricType.Precision]]
+    actual_metrics = [m for m in metrics[MetricType.Accuracy]]
     expected_metrics = [
         {
-            "type": "Precision",
-            "value": [1.0, 1.0],
+            "type": "Accuracy",
+            "value": [2 / 3, 2 / 3],
             "parameters": {
                 "score_thresholds": [0.25, 0.75],
                 "hardmax": True,
@@ -98,8 +95,8 @@ def test_precision_basic(basic_classifications: list[Classification]):
             },
         },
         {
-            "type": "Precision",
-            "value": [1.0, 0.0],
+            "type": "Accuracy",
+            "value": [1.0, 2 / 3],
             "parameters": {
                 "score_thresholds": [0.25, 0.75],
                 "hardmax": True,
@@ -113,43 +110,41 @@ def test_precision_basic(basic_classifications: list[Classification]):
         assert m in actual_metrics
 
 
-def test_precision_with_animal_example(
+def test_accuracy_with_animal_example(
     classifications_animal_example: list[Classification],
 ):
+
     loader = DataLoader()
     loader.add_data(classifications_animal_example)
     evaluator = loader.finalize()
 
-    metrics = evaluator.evaluate(
-        score_thresholds=[0.0, 0.5],
-        as_dict=True,
-    )
+    metrics = evaluator.evaluate(score_thresholds=[0.5], as_dict=True)
 
-    actual_metrics = [m for m in metrics[MetricType.Precision]]
+    actual_metrics = [m for m in metrics[MetricType.Accuracy]]
     expected_metrics = [
         {
-            "type": "Precision",
-            "value": [1.0, 1.0],
+            "type": "Accuracy",
+            "value": [2.0 / 3.0],
             "parameters": {
-                "score_thresholds": [0.0, 0.5],
+                "score_thresholds": [0.5],
                 "hardmax": True,
                 "label": "bird",
             },
         },
         {
-            "type": "Precision",
-            "value": [0.0, 0.0],
+            "type": "Accuracy",
+            "value": [0.5],
             "parameters": {
-                "score_thresholds": [0.0, 0.5],
+                "score_thresholds": [0.5],
                 "hardmax": True,
                 "label": "dog",
             },
         },
         {
-            "type": "Precision",
-            "value": [0.25, 1 / 3],
+            "type": "Accuracy",
+            "value": [2 / 3],
             "parameters": {
-                "score_thresholds": [0.0, 0.5],
+                "score_thresholds": [0.5],
                 "hardmax": True,
                 "label": "cat",
             },
@@ -161,52 +156,50 @@ def test_precision_with_animal_example(
         assert m in actual_metrics
 
 
-def test_precision_with_color_example(
+def test_accuracy_color_example(
     classifications_color_example: list[Classification],
 ):
+
     loader = DataLoader()
     loader.add_data(classifications_color_example)
     evaluator = loader.finalize()
 
-    metrics = evaluator.evaluate(
-        score_thresholds=[0.0, 0.5],
-        as_dict=True,
-    )
+    metrics = evaluator.evaluate(score_thresholds=[0.5], as_dict=True)
 
-    actual_metrics = [m for m in metrics[MetricType.Precision]]
+    actual_metrics = [m for m in metrics[MetricType.Accuracy]]
     expected_metrics = [
         {
-            "type": "Precision",
-            "value": [0.5, 0.5],
+            "type": "Accuracy",
+            "value": [2 / 3],
             "parameters": {
-                "score_thresholds": [0.0, 0.5],
+                "score_thresholds": [0.5],
                 "hardmax": True,
                 "label": "white",
             },
         },
         {
-            "type": "Precision",
-            "value": [2 / 3, 0.5],
+            "type": "Accuracy",
+            "value": [2 / 3],
             "parameters": {
-                "score_thresholds": [0.0, 0.5],
+                "score_thresholds": [0.5],
                 "hardmax": True,
                 "label": "red",
             },
         },
         {
-            "type": "Precision",
-            "value": [0.0, 0.0],
+            "type": "Accuracy",
+            "value": [2 / 3],
             "parameters": {
-                "score_thresholds": [0.0, 0.5],
+                "score_thresholds": [0.5],
                 "hardmax": True,
                 "label": "blue",
             },
         },
         {
-            "type": "Precision",
-            "value": [0.0, 0.0],
+            "type": "Accuracy",
+            "value": [5 / 6],
             "parameters": {
-                "score_thresholds": [0.0, 0.5],
+                "score_thresholds": [0.5],
                 "hardmax": True,
                 "label": "black",
             },
@@ -218,7 +211,7 @@ def test_precision_with_color_example(
         assert m in actual_metrics
 
 
-def test_precision_with_image_example(
+def test_accuracy_with_image_example(
     classifications_image_example: list[Classification],
 ):
     loader = DataLoader()
@@ -236,11 +229,11 @@ def test_precision_with_image_example(
 
     metrics = evaluator.evaluate(as_dict=True)
 
-    actual_metrics = [m for m in metrics[MetricType.Precision]]
+    actual_metrics = [m for m in metrics[MetricType.Accuracy]]
     expected_metrics = [
         {
-            "type": "Precision",
-            "value": [1.0],
+            "type": "Accuracy",
+            "value": [0.5],
             "parameters": {
                 "score_thresholds": [0.0],
                 "hardmax": True,
@@ -254,7 +247,7 @@ def test_precision_with_image_example(
         assert m in actual_metrics
 
 
-def test_precision_with_tabular_example(
+def test_accuracy_with_tabular_example(
     classifications_tabular_example: list[Classification],
 ):
     loader = DataLoader()
@@ -272,11 +265,11 @@ def test_precision_with_tabular_example(
 
     metrics = evaluator.evaluate(as_dict=True)
 
-    actual_metrics = [m for m in metrics[MetricType.Precision]]
+    actual_metrics = [m for m in metrics[MetricType.Accuracy]]
     expected_metrics = [
         {
-            "type": "Precision",
-            "value": [0.5],
+            "type": "Accuracy",
+            "value": [0.7],
             "parameters": {
                 "score_thresholds": [0.0],
                 "hardmax": True,
@@ -284,8 +277,8 @@ def test_precision_with_tabular_example(
             },
         },
         {
-            "type": "Precision",
-            "value": [0.6666666666666666],
+            "type": "Accuracy",
+            "value": [0.5],
             "parameters": {
                 "score_thresholds": [0.0],
                 "hardmax": True,
@@ -293,8 +286,8 @@ def test_precision_with_tabular_example(
             },
         },
         {
-            "type": "Precision",
-            "value": [0.0],
+            "type": "Accuracy",
+            "value": [0.8],
             "parameters": {
                 "score_thresholds": [0.0],
                 "hardmax": True,
