+++ lite/tests/classification/test_dataloader.py
@@ -1,7 +1,7 @@
 import json
 
 import pytest
-from valor_lite.cv.classification import Classification, DataLoader
+from valor_lite.classification import Classification, DataLoader
 
 
 def test_no_data():
