+++ lite/tests/classification/test_filtering.py
@@ -3,7 +3,7 @@ from random import choice, uniform
 
 import numpy as np
 import pytest
-from valor_lite.cv.classification import Classification, DataLoader, MetricType
+from valor_lite.classification import Classification, DataLoader, MetricType
 
 
 @pytest.fixture
